refactor : aws db에 맞춘 엔티티 수정작업
