--- src/main/java/com/ll/dopdang/domain/category/entity/CategoryType.java
@@ -1,10 +1,0 @@
-package com.ll.dopdang.domain.category.entity;
-
-/**
- * 카테고리 유형을 정의하는 열거형
- */
-public enum CategoryType {
-	PROJECT,
-	PROVISION,
-	PRODUCT
-}
