+++ src/main/java/com/ll/dopdang/domain/expert/category/entity/CategoryType.java
@@ -1,0 +1,7 @@
+package com.ll.dopdang.domain.expert.category.entity;
+
+public enum CategoryType {
+	PROJECT,
+	PROVISION,
+	PRODUCT
+}
