+++ src/main/java/com/ll/dopdang/domain/expert/dto/response/ExpertDetailResponseDto.java
@@ -12,13 +12,13 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExpertDetailResponseDto {
-    private Long id;              // 전문가 ID
-    private String name;          // 전문가 이름 (Member)
-    private String mainCategoryName; // 대분류 카테고리 이름
-    private List<String> subCategoryNames; // 소분류 카테고리 이름 목록
-    private String introduction; // 자기소개
-    private int careerYears;      // 경력 연수
-    private String certification; // 자격증 정보
-    private Boolean gender;       // 성별 (0 = 남자, 1 = 여자)
-    private String createdAt;     // 생성 일시
+	private Long id;              // 전문가 ID
+	private String name;          // 전문가 이름 (Member)
+	private String categoryName; // 대분류 카테고리 이름
+	private List<String> subCategoryNames; // 소분류 카테고리 이름 목록
+	private String introduction; // 자기소개
+	private int careerYears;      // 경력 연수
+	private String certification; // 자격증 정보
+	private Boolean gender;       // 성별 (0 = 남자, 1 = 여자)
+	private String createdAt;     // 생성 일시
 }
