+++ src/main/java/com/ll/dopdang/domain/expert/dto/response/ExpertResponseDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ExpertResponseDto {
 
     private String name; // 전문가 이름
-    private String mainCategoryName; // 대분류 카테고리 이름
+    private String categoryName; // 대분류 카테고리 이름
     private int careerYears; // 경력 연수
     private String introduction; // 자기소개
 }
