+++ src/main/java/com/ll/dopdang/domain/expert/repository/ExpertRepository.java
@@ -17,7 +17,7 @@ public interface ExpertRepository extends JpaRepository<Expert, Long> {
     @Query("SELECT e FROM Expert e WHERE "
             + "(e.careerYears >= :minYears OR :minYears IS NULL) AND "
             + "(e.careerYears <= :maxYears OR :maxYears IS NULL) AND "
-            + "(e.mainCategory.name IN :categoryNames OR :categoryNames IS NULL)")
+            + "(e.category.name IN :categoryNames OR :categoryNames IS NULL)")
     List<Expert> findByFilters(@Param("categoryNames") List<String> categoryNames,
                                @Param("minYears") Integer minYears,
                                @Param("maxYears") Integer maxYears);
